requirements.txt
Flask==0.10.1
Flask-Admin==1.4.0
Flask-SQLAlchemy==2.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
psycopg2==2.6.1
requests==2.9.1
SQLAlchemy==1.0.12
Werkzeug==0.11.8
WTForms==2.1
python-telegram-bot==4.0.3